move each extension to a separate file
